Refactor modals
