+++ src/components/NameCard/modal/index.js
@@ -1,0 +1,6 @@
+import { connect } from 'react-redux';
+import NameCardModal from './component';
+
+import { changeCharacterName } from '../../../actions';
+
+export default connect(null, { changeCharacterName })(NameCardModal);
